pyproject.toml
[tool.poetry]
name = "zeus"
version = "0.1.0"
description = ""
readme = "README.md"
repository = "https://github.com/getsentry/zeus"
authors = ["David Cramer <dcramer@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "~3.7"
aiodns = "~1.1.1"
aiohttp = "~3.3.2"
aioredis = "~1.1.0"
blinker = "~1.4"
requests = { version = "~2.18.4", extras = [ "security" ] }
cached-property = "~1.3.0"
cchardet = "~2.1.1"
celery = "~4.1.1"
celery-redbeat = "~0.13.0"
flask-alembic = "~2.0.1"
flask-sqlalchemy = "~2.2"
honcho = "~1.0.1"
marshmallow = "~3"
nplusone = "~0.8.1"
psycopg2-binary = "~2.7.5"
python-dateutil = "~2.6.1"
redis = "~3.2.0"
SQLAlchemy = "~1.2.0"
toronado = "~0.0.11"
unidecode = "~0.04.21"
uwsgi = "~2.0.15"
werkzeug = "~0.14.1"
cryptography = "~2.2.1"
cssselect = "~1.0.1"
cssutils = "~1.0.2"
factory_boy = "~2.12.0"
flask-mail = "~0.9.1"
flask = "~0.12.3"
google-cloud-storage = "~1.4.0"
idna = "~2.5"
ipython = "~6.1.0"
itsdangerous = "~0.24"
jsonschema = "~3.2.0"
lxml = "~4.2.3"
yarl = "~1.2.6"
requests-oauthlib = "~0.8.0"
pytest = { version = "~3.2.3", optional = true }
pytest-cov = { version = "~2.5.1", optional = true }
pytest-mock = { version = "~1.6.0", optional = true }
pytest-responses = { version = "~0.2.0", optional = true }
pytest-xdist = { version = "~1.18.0", optional = true }
responses = { version = "~0.8.1", optional = true }
sentry-sdk = { version = "~0.13.5" }

[tool.poetry.dev-dependencies]
black = "=18.6b4"
flake8 = "=3.5.0"
pycodestyle = "=2.3.1"
exam = "~0.10.6"

[tool.poetry.extras]
test = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-responses",
    "pytest-xdist",
    "responses",
]

[tool.poetry.scripts]
zeus = 'zeus.cli:main'

[build-system]
requires = ["poetry>=0.12.7", "pip==19.3.1"]
build-backend = "poetry.masonry.api"
